Añadir endpoint Materia en rúbrica
Se agregó un endpoint para consultar todas las materias guardadas en BD. Implicó crear en la capa de dominio una clase de Materia y en el mapper de rúbrica un método que permita Mapear las materias de Materia a MateriaDTORespuesta.

Se cambió el crfPorcentaje de float a Integer. Implicó cambiar los constructores.

+++ backend/ModuloRubricaCriterio_Hexagonal/src/main/java/org/unicauca/modulorubricacriterio/Dominio/Modelos/Criterio.java
@@ -12,7 +12,7 @@ public class Criterio {
     private Long idCriterio;
     private Rubrica rubrica;
     private String crfDescripcion;
-    private Float crfPorcentaje;
+    private Integer crfPorcentaje;
     private Float crfNota;
     private String crfComentario;
     private List<Nivel> niveles;
